map markers for product and user locations done

+++ app/src/main/java/com/example/lab_1_2_shubhambehal_c0835488_android/interfaces/ProductEditDeleteListener.java
@@ -5,5 +5,5 @@ public interface ProductEditDeleteListener {
 
     void editProduct(int productId);
 
-    void openMap(int latitude, int longitude);
+    void openMap(double latitude, double longitude);
 }
